% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parglm-package.R
\docType{package}
\name{parglm-package}
\alias{parglm-package}
\title{parglm: Parallel GLM}
\description{
Provides a parallel estimation method for generalized linear models without compiling with a multithreaded LAPACK or BLAS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/remlapmot/parglm}
  \item \url{https://remlapmot.github.io/parglm/}
  \item Report bugs at \url{https://github.com/remlapmot/parglm/issues}
}

}
\author{
\strong{Maintainer}: Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})

Authors:
\itemize{
  \item Tom Palmer \email{remlapmot@hotmail.com} (\href{https://orcid.org/0000-0003-4655-4511}{ORCID})
  \item Benjamin Christoffersen \email{boennecd@gmail.com} (\href{https://orcid.org/0000-0002-7182-1346}{ORCID})
}

Other contributors:
\itemize{
  \item Anthony Williams [copyright holder]
  \item Boost developers [copyright holder]
}

}
\keyword{internal}
